+++ src/sound/audio_stream/msg/audiodata.msg
@@ -1,2 +1,3 @@
 float32 Decibel
 float32 Frequency
+float32 suddenchange
